Evaluated LVP on 6 benchmarks on Atomic

+++ src/cpu/vpred/lvp.hh
@@ -20,7 +20,7 @@ class LVP : public VPredUnit
         LVP(const LVPParams *params);
 
         bool lookup(Addr inst_addr, RegVal &value);
-        float getconf(Addr inst_addr, RegVal &value);
+        uint8_t getconf(Addr inst_addr, RegVal &value);
 
         void updateTable(Addr inst_addr, bool isValuePredicted, bool isValueTaken, RegVal &trueValue);
